### 2025-06-27 #713 @atsyplenkov
tidyhydro — presubmission inquiry
Submitting Author Name: Anatoly Tsyplenkov
Submitting Author Github Handle: @atsyplenkov
Repository:  https://github.com/atsyplenkov/tidyhydro
Submission type: Pre-submission

---

```
Package: tidyhydro
Title: Tidy Metrics for Assessing Hydrological Models Performance
Version: 0.1.1
    person(given = "Anatoly", family = "Tsyplenkov", email = "atsyplenkov@fastmail.com", role = c("cre", "aut", "cph"), comment = c(ORCID = "0000-0003-4144-8402"))
Maintainer: Anatoly Tsyplenkov <atsyplenkov@fastmail.com>
Description: Provides tidy tools for comparing simulated and observed hydrological time series. Includes compatibility with the 'yardstick' package for model performance evaluation using commonly used metrics such as the Nash–Sutcliffe Efficiency (NSE), Kling–Gupta Efficiency (KGE), percent bias (pBIAS) and etc.
    Rcpp (>= 1.0.12),
    rlang (>= 1.1.0),
    yardstick (>= 1.3.1)
LinkingTo: Rcpp
    hydroGOF,
    quickcheck (>= 0.1.3),
    quarto
Config/Needs/website: bench, ggplot2, quarto, lubridate, dplyr
```

`tidyhydro` fits within the data validation and testing category / regression , as it provides automated, quantitative metrics (e.g., NSE, KGE, pBIAS) for evaluating the quality and performance of hydrological models compatible with the `tidymodels` infrastructure.

No, I just learned about the `srr` from the issue template. Feel a bit confused regarding that standards and not sure if it is applicable to my package.

The target audience comprises the hydrological/ecological community and the general data science community who uses regression models within their workflows.

Yes, there is a package [`hydroGOF`](https://github.com/hzambran/hydroGOF), which has some overlap with `tidyhydro` in terms of functionality. For example, metrics such as MSE, pBIAS, NSE, and KGE are implemented in both packages. However, `tidyhydro` provides a `tidyverse`-style interface, giving full compatibility and interoperability with the popular `tidymodels` framework (via `yardstick`). Additionally, `tidyhydro` uses C++ via Rcpp for performance, which helps in validating large time series efficiently (see [benchmarks](https://atsyplenkov.github.io/tidyhydro/articles/benchmarks.html)).

1. I raised this issue out of curiosity, as I have been thinking about submitting some package to rOpenSci for several years. However, I am not 100% sure that `tidyhydro` falls within the aims and scope of rOpenSci. I am interested to hear your thoughts about it.
2. The local `pkgcheck` fails to find continuous integration checks, which confuses me, as I do have a very comprehensive CI workflow set up. I am curious to see the results of the rOpenSci-hosted `pkgcheck`.
3. I do need some guidance regarding `srr` standards. As far as I can understand from the `srr` manual, it is not really applicable to my package, as `tidyhydro` is not introducing new methods but rather provides a re-implementation of a well-known ones in a tidy way.
