### 2025-06-27 #713 @mpadge
Thanks @atsyplenkov for your inquiry here. My first impressions are that the package seems like a good fit here indeed! Short responses to your specific questions:

>     1. I raised this issue out of curiosity, as I have been thinking about submitting some package to rOpenSci for several years. However, I am not 100% sure that `tidyhydro` falls within the aims and scope of rOpenSci. I am interested to hear your thoughts about it.

Already stated above: I'm very confident that this package will be a good fit (with proviso under 3rd point below).

>     2. The local `pkgcheck` fails to find continuous integration checks, which confuses me, as I do have a very comprehensive CI workflow set up. I am curious to see the results of the rOpenSci-hosted `pkgcheck`.

Sorry for confusion there. That confusing message has been [updated this week](https://github.com/ropensci-review-tools/pkgcheck/issues/255). It won't be deployed in our bot here until next week, but you can also install https://github.com/ropensci-review-tools/pkgcheck-action in your repo to generate reports on every push.

>     3. I do need some guidance regarding `srr` standards. As far as I can understand from the `srr` manual, it is not really applicable to my package, as `tidyhydro` is not introducing new methods but rather provides a re-implementation of a well-known ones in a tidy way.

It's perfectly okay for a stats package to be a re-implementation. I think the biggest initial challenge we'll have here is deciding in which category the package best fits. For that, the guideline is that any category in which you think you'll be able to comply with at least half of standards is the right category. You've indicated _Regression and Supervised Learning_ above, although it seems to me that the actual implementations are more summary estimators than regression models. Can you please have a look through [our standards](https://stats-devguide.ropensci.org/standards.html) and think about how well you'll be able to meet these categories:

- [_Regression and Supervised Learning_](https://stats-devguide.ropensci.org/standards.html#standards-regression), because that's what you indicated above
- [_Exploratory Data Analysis and Summary Statistics_](https://stats-devguide.ropensci.org/standards.html#standards-eda) which I think may be a good fit
- [_Time Series_](https://stats-devguide.ropensci.org/standards.html#standards-time-series) may be applicable, because as far as I understand your package, initial input data are presumed to be hydrological time series.
- [_Probability Distributions_](https://stats-devguide.ropensci.org/standards.html#standards-distributions) which I do not think will apply here, but I suspect just reading those may give you a different perspective on some of the implications of distributional assumptions currently hard-coded within your routines.

It's okay to tick multiple categories, but that just entails more work for you in documenting compliance standards. In short: Please look through, and report back on which category you think you'll be able to comply with the greatest number of standards. Thanks, and feel free to ask any questions here in the meantime :+1:
