### 2025-06-27 #713 @atsyplenkov
Hi @mpadge 👋! I appreciate your quick and very detailed response. It has definitely shed a lot of light! I agree that the *Exploratory Data Analysis and Summary Statistics* category feels like the best fit. This is evident from the examples listed on the website.

That being said, I am happy to proceed with the full package review in the near future. I just need some time to check the documentation against the `srr` standards, as I wish to aim at the silver badge at least. I will raise an issue when I am ready.

One more question regarding issue templates. There is a significant difference between general software and statistical software. I noticed that the template for statistical packages doesn't include a MEE option. Does this mean statistical software is not welcome in MEE, or is the template just outdated?
