### 2025-06-27 6654ab4
ci: pkgcheck workflow
