+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidyhydro (development version)
+
 # tidyhydro 0.1.1
 
 ## New features
