### 2025-06-28 f0d648e
docs: update CONTRIBUTING
