### 2025-06-28 51307fd
docs: add keywords to references
