+++ man/nse.Rd
@@ -118,3 +118,4 @@ Other accuracy metrics:
 }
 \concept{accuracy metrics}
 \concept{numeric metrics}
+\keyword{gof}
