+++ man/sfe.Rd
@@ -104,3 +104,4 @@ Other accuracy metrics:
 }
 \concept{accuracy metrics}
 \concept{numeric metrics}
+\keyword{regression}
