+++ man/kge2012.Rd
@@ -126,3 +126,4 @@ Other accuracy metrics:
 }
 \concept{accuracy metrics}
 \concept{numeric metrics}
+\keyword{gof}
