+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   articles/benchmarks: benchmarks.html
   articles/tidyhydro: tidyhydro.html
-last_built: 2025-06-28T00:00Z
+last_built: 2025-06-28T04:11Z
 urls:
   reference: https://atsyplenkov.github.io/tidyhydro/reference
   article: https://atsyplenkov.github.io/tidyhydro/articles
