### 2025-06-28 c62045a
ci: fix typos in workflows
