+++ .github/workflows/pkgcheck.yaml
@@ -8,11 +8,7 @@ concurrency:
 on:
   # Manually trigger the Action under Actions/pkgcheck
   workflow_dispatch:
-  # Run on every push to main
-  push:
-    branches:
-      - main
-      - master
+
 
 jobs:
   pkgcheck:
