+++ .github/workflows/check-r-pkg.yaml
@@ -2,7 +2,7 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: master]
+    branches: master
   pull_request:
     branches: master
 
