### 2025-06-28 4e94b4c
feat: add RMSE
