+++ man/rmse.Rd
@@ -1,0 +1,89 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mse.R
+\name{rmse}
+\alias{rmse}
+\alias{rmse.data.frame}
+\alias{rmse_vec}
+\title{Root Mean Squared Error (RMSE)}
+\usage{
+rmse(data, ...)
+
+\method{rmse}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)
+
+rmse_vec(truth, estimate, na_rm = TRUE, ...)
+}
+\arguments{
+\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
+and \code{estimate} arguments.}
+
+\item{...}{Not currently used.}
+
+\item{truth}{The column identifier for the true results
+(that is \code{numeric}). This should be an unquoted column name although
+this argument is passed by expression and supports
+\link[rlang:topic-inject]{quasiquotation} (you can unquote column
+names). For \verb{_vec()} functions, a \code{numeric} vector.}
+
+\item{estimate}{The column identifier for the predicted
+results (that is also \code{numeric}). As with \code{truth} this can be
+specified different ways but the primary method is to use an
+unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}
+
+\item{na_rm}{A \code{logical} value indicating whether \code{NA}
+values should be stripped before the computation proceeds.}
+}
+\value{
+A \code{tibble} with columns \code{.metric}, \code{.estimator},
+and \code{.estimate} and 1 row of values.
+
+For grouped data frames, the number of rows returned will be the same as
+the number of groups.
+
+For \code{rmse_vec()}, a single \code{numeric} value (or \code{NA}).
+}
+\description{
+Root Mean Squared Error (RMSE)
+}
+\details{
+The RMSE is estimated as follows:
+\deqn{
+RMSE = \sqrt{\frac{1}{n} \sum_{i=1}^{n}{(sim_i - obs_i)^2}}
+}
+where:
+\itemize{
+\item \eqn{sim} defines model simulations at time step \eqn{i}
+\item \eqn{obs} defines model observations at time step \eqn{i}
+}
+}
+\examples{
+library(tidyhydro)
+data(avacha)
+
+# Supply truth and predictions as bare column names
+rmse(avacha, obs, sim)
+
+# Or as numeric vectors
+rmse_vec(avacha$obs, avacha$sim)
+}
+\seealso{
+Other numeric metrics: 
+\code{\link{kge}()},
+\code{\link{kge2012}()},
+\code{\link{mse}()},
+\code{\link{nse}()},
+\code{\link{pbias}()},
+\code{\link{press}()},
+\code{\link{sfe}()}
+
+Other accuracy metrics: 
+\code{\link{kge}()},
+\code{\link{kge2012}()},
+\code{\link{mse}()},
+\code{\link{nse}()},
+\code{\link{pbias}()},
+\code{\link{press}()},
+\code{\link{sfe}()}
+}
+\concept{accuracy metrics}
+\concept{numeric metrics}
+\keyword{gof}
