+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(nse, data.frame)
 S3method(kge, data.frame)
 S3method(kge2012, data.frame)
 S3method(mse, data.frame)
+S3method(rmse, data.frame)
 S3method(pbias, data.frame)
 S3method(press, data.frame)
 S3method(sfe, data.frame)
@@ -13,6 +14,7 @@ export(nse)
 export(kge)
 export(kge2012)
 export(mse)
+export(rmse)
 export(pbias)
 export(press)
 export(sfe)
@@ -20,6 +22,7 @@ export(nse_vec)
 export(kge_vec)
 export(kge2012_vec)
 export(mse_vec)
+export(rmse_vec)
 export(pbias_vec)
 export(press_vec)
 export(sfe_vec)
