### 2025-06-28 ff1f568
docs: spelling and grammar
