+++ README.md
@@ -26,15 +26,15 @@ library is mainly written in C++ and provides a very quick estimation of
 desired goodness-of-fit criteria.
 
 Additionally, you’ll find here a C++ implementation of lesser-known yet
-powerful metrics used in reports from the United States Geological
-Survey (USGS) and the National Environmental Monitoring Standards (NEMS)
-guidelines. Examples include *PRESS* (Prediction Error Sum of Squares),
-*SFE* (Standard Factorial Error), and *MSPE* (Model Standard Percentage
-Error) and others. Based on the equations from *Helsel et al.*
-([2020](https://pubs.usgs.gov/publication/tm4A3)), *Rasmunsen et al.*
-([2008](https://pubs.usgs.gov/tm/tm3c4/)), *Hicks et al.*
-([2020](https://www.nems.org.nz/documents/suspended-sediment)) and etc.
-(see documentation for details).
+powerful metrics and coefficients recommended in the United States
+Geological Survey (USGS) and the National Environmental Monitoring
+Standards (NEMS) guidelines. Examples include *PRESS* (Prediction Error
+Sum of Squares), *SFE* (Standard Factorial Error), *MSPE* (Model
+Standard Percentage Error) and others. Based on the equations from
+*Helsel et al.* ([2020](https://pubs.usgs.gov/publication/tm4A3)),
+*Rasmunsen et al.* ([2008](https://pubs.usgs.gov/tm/tm3c4/)), *Hicks et
+al.* ([2020](https://www.nems.org.nz/documents/suspended-sediment)) and
+etc. (see documentation for details).
 
 ## Example
 
@@ -128,9 +128,9 @@ bench::mark(
 #> # A tibble: 3 × 6
 #>   expression   min median `itr/sec` mem_alloc `gc/sec`
 #>   <bch:expr> <dbl>  <dbl>     <dbl>     <dbl>    <dbl>
-#> 1 tidyhydro   1       1       22.7        NaN      NaN
-#> 2 hydroGOF   15.2    19.1      1          Inf      Inf
-#> 3 baseR       8.66   10.6      2.44       Inf      Inf
+#> 1 tidyhydro   1       1       30.2        NaN      NaN
+#> 2 hydroGOF   14.8    20.0      1          Inf      Inf
+#> 3 baseR       8.44   10.3      2.55       Inf      Inf
 ```
 
 ## Code of Conduct
