### 2025-06-28 b38bb2d
fix: removed quarto from soft deps
