+++ DESCRIPTION
@@ -20,8 +20,7 @@ Language: en-US
 Suggests:
     hydroGOF,
     testthat (>= 3.0.0),
-    quickcheck (>= 0.1.3),
-    quarto
+    quickcheck (>= 0.1.3)
 Config/testthat/edition: 3
 URL: https://github.com/atsyplenkov/tidyhydro, https://atsyplenkov.github.io/tidyhydro/
 BugReports: https://github.com/atsyplenkov/tidyhydro/issues
