+++ tests/testthat/test-hydrogof.R
@@ -16,14 +16,14 @@ test_that("nse", {
   # same results as hydroGOF with both NA values present and without
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -38,14 +38,14 @@ test_that("nse", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
@@ -65,14 +65,14 @@ test_that("kge", {
   # With NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -87,14 +87,14 @@ test_that("kge", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
@@ -115,14 +115,14 @@ test_that("kge2012", {
   # With NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -142,14 +142,14 @@ test_that("kge2012", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
@@ -174,14 +174,14 @@ test_that("pbias", {
   # With NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -196,14 +196,14 @@ test_that("pbias", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
[…]
