### 2025-07-02 #713 @mpadge
Thanks @atsyplenkov, i'll put this issue on hold for the moment then. That just means the bot will add a "holding" label to this issue, which we can remove once things get going again. When you're ready, we'll remove that label, run our automated checks here, and hopefully proceed to a full submission from that point. The bot will send a reminder after 3 months.

As for MEE: That's a definite Yes, just that your package would be the first stats package to go on to MEE. Our understanding with them is that our peer review is acceptable as a phase of review for MEE, and there's no reason why stats peer review would be considered any different (especially given that the requirements are considerably more stringent than for non-stats review). If you were to follow that track, then we'd request your help in updating our templates and associated documentation once things get that far.
