+++ man/kge.Rd
@@ -105,6 +105,7 @@ Kling–Gupta efficiency scores. Hydrology and Earth System Sciences, 23,
 }
 \seealso{
 Other numeric metrics: 
+\code{\link{cv}()},
 \code{\link{kge2012}()},
 \code{\link{mse}()},
 \code{\link{nse}()},
@@ -114,6 +115,7 @@ Other numeric metrics:
 \code{\link{sfe}()}
 
 Other accuracy metrics: 
+\code{\link{cv}()},
 \code{\link{kge2012}()},
 \code{\link{mse}()},
 \code{\link{nse}()},
