+++ man/cv.Rd
@@ -1,16 +1,16 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/mse.R
-\name{rmse}
-\alias{rmse}
-\alias{rmse.data.frame}
-\alias{rmse_vec}
-\title{Root Mean Squared Error (RMSE)}
+% Please edit documentation in R/variability.R
+\name{cv}
+\alias{cv}
+\alias{cv.data.frame}
+\alias{cv_vec}
+\title{Coefficient of Variation (Cv)}
 \usage{
-rmse(data, ...)
+cv(data, ...)
 
-\method{rmse}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)
+\method{cv}{data.frame}(data, truth, na_rm = TRUE, ...)
 
-rmse_vec(truth, estimate, na_rm = TRUE, ...)
+cv_vec(truth, na_rm = TRUE, ...)
 }
 \arguments{
 \item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
@@ -24,11 +24,6 @@ this argument is passed by expression and supports
 \link[rlang:topic-inject]{quasiquotation} (you can unquote column
 names). For \verb{_vec()} functions, a \code{numeric} vector.}
 
-\item{estimate}{The column identifier for the predicted
-results (that is also \code{numeric}). As with \code{truth} this can be
-specified different ways but the primary method is to use an
-unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}
-
 \item{na_rm}{A \code{logical} value indicating whether \code{NA}
 values should be stripped before the computation proceeds.}
 }
@@ -39,53 +34,42 @@ and \code{.estimate} and 1 row of values.
 For grouped data frames, the number of rows returned will be the same as
 the number of groups.
 
-For \code{rmse_vec()}, a single \code{numeric} value (or \code{NA}).
+For \code{cv_vec()}, a single \code{numeric} value (or \code{NA}).
 }
 \description{
-Root Mean Squared Error (RMSE)
-}
-\details{
-The RMSE is estimated as follows:
-\deqn{
-RMSE = \sqrt{\frac{1}{n} \sum_{i=1}^{n}{(sim_i - obs_i)^2}}
-}
-where:
-\itemize{
-\item \eqn{sim} defines model simulations at time step \eqn{i}
-\item \eqn{obs} defines model observations at time step \eqn{i}
-}
+Coefficient of Variation (Cv)
 }
 \examples{
 library(tidyhydro)
 data(avacha)
 
 # Supply truth and predictions as bare column names
-rmse(avacha, obs, sim)
+cv(avacha, obs)
 
 # Or as numeric vectors
-rmse_vec(avacha$obs, avacha$sim)
+cv_vec(avacha$obs)
 }
 \seealso{
 Other numeric metrics: 
-\code{\link{cv}()},
 \code{\link{kge}()},
 \code{\link{kge2012}()},
 \code{\link{mse}()},
 \code{\link{nse}()},
 \code{\link{pbias}()},
 \code{\link{press}()},
+\code{\link{rmse}()},
 \code{\link{sfe}()}
 
 Other accuracy metrics: 
-\code{\link{cv}()},
 \code{\link{kge}()},
 \code{\link{kge2012}()},
 \code{\link{mse}()},
 \code{\link{nse}()},
 \code{\link{pbias}()},
 \code{\link{press}()},
+\code{\link{rmse}()},
 \code{\link{sfe}()}
 }
 \concept{accuracy metrics}
 \concept{numeric metrics}
-\keyword{gof}
+\keyword{summary}

[…]
