### 2025-07-09 76e3d51
feat: add `measure` class
