+++ man-roxygen/examples-description.R
@@ -1,8 +1,9 @@
 #' @examples
 #' library(tidyhydro)
-#' data(avacha)
 #'
-#' # Supply truth and predictions as bare column names
+#' <%=fn %>
+#' 
+#' # Supply truth as bare column names
 #' <%=fn %>(avacha, obs)
 #'
 #' # Or as numeric vectors
