+++ man/new-measure.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aaa-new.R
+\name{new-measure}
+\alias{new-measure}
+\alias{new_tendency_measure}
+\alias{new_var_measure}
+\alias{new_sym_measure}
+\title{Construct a new measure function}
+\usage{
+new_tendency_measure(fn)
+
+new_var_measure(fn)
+
+new_sym_measure(fn)
+}
+\arguments{
+\item{fn}{A function. The measure function to attach a measure-specific class}
+}
+\description{
+These functions provide convenient wrappers to create the three types of
+measure functions in \code{tidyhydro}: measures of central tendency, variability
+and symmetry. They add a measure-specific class to \code{fn} and
+mimic a behaviour of \code{\link[yardstick:metric_set]{yardstick::metric_set()}}. These features are used
+by \code{\link[=measure_set]{measure_set()}}.
+
+See \href{https://www.tidymodels.org/learn/develop/metrics/}{Custom performance metrics} for more
+information about creating custom metrics.
+}
+\keyword{summary_stats}

[…]
