### 2025-07-09 5873c17
lint: making rlang calls explicit
