+++ R/import-standalone-obj-type.R
@@ -67,7 +67,7 @@
 #'   article, e.g. "an integer vector".
 #' @noRd
 obj_type_friendly <- function(x, value = TRUE) {
-  if (is_missing(x)) {
+  if (rlang::is_missing(x)) {
     return("absent")
   }
 
@@ -80,15 +80,15 @@ obj_type_friendly <- function(x, value = TRUE) {
     return(sprintf("a <%s> object", type))
   }
 
-  if (!is_vector(x)) {
+  if (!rlang::is_vector(x)) {
     return(.rlang_as_friendly_type(typeof(x)))
   }
 
   n_dim <- length(dim(x))
 
   if (!n_dim) {
-    if (!is_list(x) && length(x) == 1) {
-      if (is_na(x)) {
+    if (!rlang::is_list(x) && length(x) == 1) {
+      if (rlang::is_na(x)) {
         return(switch(
           typeof(x),
           logical = "`NA`",
@@ -173,8 +173,8 @@ obj_type_friendly <- function(x, value = TRUE) {
 }
 
 vec_type_friendly <- function(x, length = FALSE) {
-  if (!is_vector(x)) {
-    abort("`x` must be a vector.")
+  if (!rlang::is_vector(x)) {
+    rlang::abort("`x` must be a vector.")
   }
   type <- typeof(x)
   n_dim <- length(dim(x))
@@ -260,8 +260,8 @@ vec_type_friendly <- function(x, length = FALSE) {
   )
 }
 
-.rlang_stop_unexpected_typeof <- function(x, call = caller_env()) {
-  abort(
+.rlang_stop_unexpected_typeof <- function(x, call = rlang::caller_env()) {
+  rlang::abort(
     sprintf("Unexpected type <%s>.", typeof(x)),
     call = call
   )
@@ -296,7 +296,7 @@ obj_type_oo <- function(x) {
 #'   character vector of expected types, in which case the error
 #'   message mentions all of them in an "or" enumeration.
 #' @param show_value Passed to `value` argument of `obj_type_friendly()`.
-#' @param ... Arguments passed to [abort()].
+#' @param ... Arguments passed to [rlang::abort()].
 #' @inheritParams args_error_context
 #' @noRd
 stop_input_type <- function(
@@ -306,11 +306,11 @@ stop_input_type <- function(
   allow_na = FALSE,
   allow_null = FALSE,
   show_value = TRUE,
-  arg = caller_arg(x),
-  call = caller_env()
+  arg = rlang::caller_arg(x),
+  call = rlang::caller_env()
 ) {
   # From standalone-cli.R
-  cli <- env_get_list(
+  cli <- rlang::env_get_list(
     nms = c("format_arg", "format_code"),
     last = topenv(),
     default = function(x) sprintf("`%s`", x),
@@ -339,7 +339,7 @@ stop_input_type <- function(
     obj_type_friendly(x, value = show_value)
   )
 
-  abort(message, ..., call = call, arg = arg)
+  rlang::abort(message, ..., call = call, arg = arg)
 }
 
 oxford_comma <- function(chr, sep = ", ", final = "or") {
