### 2025-07-09 eba2c8f
docs: update new-measure description
