### 2025-07-09 d4c13df
fix: proper rlang imports and types check
