+++ R/import-standalone-types-check.R
@@ -70,8 +70,8 @@ check_bool <- function(
   ...,
   allow_na = FALSE,
   allow_null = FALSE,
-  arg = rlang::caller_arg(x),
-  call = rlang::caller_env()
+  arg = caller_arg(x),
+  call = caller_env()
 ) {
   if (
     !missing(x) &&
@@ -102,8 +102,8 @@ check_string <- function(
   allow_empty = TRUE,
   allow_na = FALSE,
   allow_null = FALSE,
-  arg = rlang::caller_arg(x),
-  call = rlang::caller_env()
+  arg = caller_arg(x),
+  call = caller_env()
 ) {
   if (!missing(x)) {
     is_string <- .rlang_check_is_string(
@@ -129,13 +129,13 @@ check_string <- function(
 }
 
 .rlang_check_is_string <- function(x, allow_empty, allow_na, allow_null) {
-  if (rlang::is_string(x)) {
-    if (allow_empty || !rlang::is_string(x, "")) {
+  if (is_string(x)) {
+    if (allow_empty || !is_string(x, "")) {
       return(TRUE)
     }
   }
 
-  if (allow_null && rlang::is_null(x)) {
+  if (allow_null && is_null(x)) {
     return(TRUE)
   }
 
@@ -150,8 +150,8 @@ check_name <- function(
   x,
   ...,
   allow_null = FALSE,
-  arg = rlang::caller_arg(x),
-  call = rlang::caller_env()
+  arg = caller_arg(x),
+  call = caller_env()
 ) {
   if (!missing(x)) {
     is_string <- .rlang_check_is_string(
@@ -188,8 +188,8 @@ check_number_decimal <- function(
   allow_infinite = TRUE,
   allow_na = FALSE,
   allow_null = FALSE,
-  arg = rlang::caller_arg(x),
-  call = rlang::caller_env()
+  arg = caller_arg(x),
+  call = caller_env()
 ) {
   if (missing(x)) {
     exit_code <- IS_NUMBER_false
@@ -231,8 +231,8 @@ check_number_whole <- function(
   allow_infinite = FALSE,
   allow_na = FALSE,
   allow_null = FALSE,
-  arg = rlang::caller_arg(x),
-  call = rlang::caller_env()
+  arg = caller_arg(x),
+  call = caller_env()
 ) {
   if (missing(x)) {
     exit_code <- IS_NUMBER_false
@@ -295,7 +295,7 @@ check_number_whole <- function(
     } else if (x > max) {
       what <- sprintf("%s smaller than or equal to %s", what, max)
     } else {
-      rlang::abort("Unexpected state in OOB check", .internal = TRUE)
+      abort("Unexpected state in OOB check", .internal = TRUE)
     }
   }
 
@@ -314,14 +314,14 @@ check_symbol <- function(
   x,
   ...,
   allow_null = FALSE,
-  arg = rlang::caller_arg(x),
-  call = rlang::caller_env()
+  arg = caller_arg(x),
+  call = caller_env()
 ) {
   if (!missing(x)) {
-    if (rlang::is_symbol(x)) {
+    if (is_symbol(x)) {
       return(invisible(NULL))
     }
-    if (allow_null && rlang::is_null(x)) {
+    if (allow_null && is_null(x)) {
       return(invisible(NULL))
     }
   }
@@ -341,14 +341,14 @@ check_arg <- function(
   x,
   ...,
   allow_null = FALSE,
-  arg = rlang::caller_arg(x),
-  call = rlang::caller_env()
+  arg = caller_arg(x),
+  call = caller_env()
 ) {
   if (!missing(x)) {
-    if (rlang::is_symbol(x)) {
+    if (is_symbol(x)) {
       return(invisible(NULL))
     }
[…]
