+++ DESCRIPTION
@@ -11,7 +11,8 @@ Depends: R (>= 4.1.0)
 Imports:
     Rcpp (>= 1.0.12),
     rlang (>= 1.1.0),
-    yardstick (>= 1.3.1)
+    yardstick (>= 1.3.1),
+    cli
 LinkingTo: Rcpp
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
