### 2025-07-09 6fa04c5
feat: geometric mean
