+++ NAMESPACE
@@ -1,19 +1,11 @@
 useDynLib(tidyhydro, .registration = TRUE)
 importFrom(Rcpp, evalCpp)
-import(rlang)
 # exportPattern("^[[:alpha:]]+")
 
-S3method(nse, data.frame)
-S3method(kge, data.frame)
-S3method(kge2012, data.frame)
-S3method(mse, data.frame)
-S3method(rmse, data.frame)
-S3method(pbias, data.frame)
-S3method(press, data.frame)
-S3method(cv, data.frame)
-S3method(sfe, data.frame)
 S3method(print, measure)
 S3method(format, measure)
+
+# general functions
 export(nse)
 export(kge)
 export(kge2012)
@@ -23,6 +15,21 @@ export(pbias)
 export(press)
 export(sfe)
 export(cv)
+export(gm)
+
+# data.frame methods
+S3method(nse, data.frame)
+S3method(kge, data.frame)
+S3method(kge2012, data.frame)
+S3method(mse, data.frame)
+S3method(rmse, data.frame)
+S3method(pbias, data.frame)
+S3method(press, data.frame)
+S3method(sfe, data.frame)
+S3method(cv, data.frame)
+S3method(gm, data.frame)
+
+# vector functions
 export(nse_vec)
 export(kge_vec)
 export(kge2012_vec)
@@ -32,3 +39,4 @@ export(pbias_vec)
 export(press_vec)
 export(sfe_vec)
 export(cv_vec)
+export(gm_vec)
