### 2025-07-10 27ad61d
added paper skeleton
