+++ paper/references.bib
@@ -1,0 +1,344 @@
+@article{aertsLargesampleAssessmentVarying2022,
+  title = {Large-Sample Assessment of Varying Spatial Resolution on the Streamflow Estimates of the Wflow\_sbm Hydrological Model},
+  author = {Aerts, Jerom P. M. and Hut, Rolf W. and {van de Giesen}, Nick C. and Drost, Niels and {van Verseveld}, Willem J. and Weerts, Albrecht H. and Hazenberg, Pieter},
+  year = {2022},
+  month = aug,
+  journal = {Hydrology and Earth System Sciences},
+  volume = {26},
+  number = {16},
+  pages = {4407--4430},
+  publisher = {Copernicus GmbH},
+  issn = {1027-5606},
+  doi = {10.5194/hess-26-4407-2022},
+  url = {https://hess.copernicus.org/articles/26/4407/2022/},
+  urldate = {2025-06-24},
+  abstract = {Distributed hydrological modelling moves into the realm of hyper-resolution modelling. This results in a plethora of scaling-related challenges that remain unsolved. To the user, in light of model result interpretation, finer-resolution output might imply an increase in understanding of the complex interplay of heterogeneity within the hydrological system. Here we investigate spatial scaling in the form of varying spatial resolution by evaluating the streamflow estimates of the distributed wflow\_sbm hydrological model based on 454 basins from the large-sample CAMELS data set. Model instances are derived at three spatial resolutions, namely 3 km, 1 km, and 200 m. The results show that a finer spatial resolution does not necessarily lead to better streamflow estimates at the basin outlet. Statistical testing of the objective function distributions (Kling--Gupta efficiency (KGE) score) of the three model instances resulted in only a statistical difference between the 3 km and 200 m streamflow estimates. However, an assessment of sampling uncertainty shows high uncertainties surrounding the KGE score throughout the domain. This makes the conclusion based on the statistical testing inconclusive. The results do indicate strong locality in the differences between model instances expressed by differences in KGE scores of on average 0.22 with values larger than 0.5. The results of this study open up research paths that can investigate the changes in flux and state partitioning due to spatial scaling. This will help to further understand the challenges that need to be resolved for hyper-resolution hydrological modelling.},
+  langid = {english},
+  file = {C:\Users\TsyplenkovA\OneDrive - MWLR\ATS\Personal\zotero-library\Aerts et al.\Aerts et al._2022_Large-sample assessment of varying spatial resolut.pdf}
+}
+
+@article{clarkAbusePopularPerformance2021,
+  title = {The {{Abuse}} of {{Popular Performance Metrics}} in {{Hydrologic Modeling}}},
+  author = {Clark, Martyn P. and Vogel, Richard M. and Lamontagne, Jonathan R. and Mizukami, Naoki and Knoben, Wouter J. M. and Tang, Guoqiang and Gharari, Shervan and Freer, Jim E. and Whitfield, Paul H. and Shook, Kevin R. and Papalexiou, Simon Michael},
+  year = {2021},
+  journal = {Water Resources Research},
+  volume = {57},
+  number = {9},
+  pages = {e2020WR029001},
+  issn = {1944-7973},
+  doi = {10.1029/2020WR029001},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1029/2020WR029001},
+  urldate = {2025-06-22},
[…]
