+++ .github/CONTRIBUTING.md
@@ -50,7 +50,9 @@ Please note that the tidyhydro project is released with a [Contributor Code of C
 ### GOF
 -[ ] KGE''
 
--[ ] KGE Score
+-[ ] KGE Score [@knobenTechnicalNoteInherent2019]
+
+-[ ] Weighted KGE [@mizukamiChoiceCalibrationMetrics2019]
 
 -[ ] MSDR - mean squared deviation ratio [@oliverTutorialGuideGeostatistics2014]
 
@@ -67,3 +69,13 @@ Please note that the tidyhydro project is released with a [Contributor Code of C
 
 -[ ] MSPE - model standard percentage error log and nonlog [@rasmussenGuidelinesProceduresComputing2009, p. 13]
 
+### Stats
+-[ ] MAD
+
+-[x] Geometric Mean (GM)
+
+-[x] Coefficient of Variation (Cv)
+
+-[ ] Coefficient of Skewness (g or Cs)
+
+-[ ] Others measures of descriptive stats mentioned in Helsel et al. (2020)
