+++ paper/references.bib
@@ -221,6 +221,24 @@
   file = {C:\Users\TsyplenkovA\OneDrive - MWLR\ATS\Personal\zotero-library\Markowska et al.\Markowska et al._2025_rangr An R package for mechanistic, spatially exp.pdf}
 }
 
+@article{melsenRiseNashSutcliffeEfficiency2025,
+  title = {The Rise of the {{Nash-Sutcliffe}} Efficiency in Hydrology},
+  author = {Melsen, Lieke A. and Puy, Arnald and Torfs, Paul J. J. F. and Saltelli, Andrea},
+  year = {2025},
+  month = jun,
+  journal = {Hydrological Sciences Journal},
+  volume = {70},
+  number = {8},
+  pages = {1248--1259},
+  publisher = {Taylor \& Francis},
+  issn = {0262-6667},
+  doi = {10.1080/02626667.2025.2475105},
+  url = {https://doi.org/10.1080/02626667.2025.2475105},
+  urldate = {2025-07-12},
+  abstract = {The Nash-Sutcliffe efficiency (NSE) is commonly used as a model evaluation metric in hydrology, but its prominence is often taken for granted. This study explores the social factors behind its adoption. Introduced in 1970, the NSE gained traction as computational advancements spurred the growth of hydrological models and evaluation metrics. This, in turn, led to the need to converge on broadly accepted metrics. In 1990, a committee recommended the NSE alongside two other metrics. One of the main developers of SWAT, a widely used hydrological model, adopted only the NSE part of this recommendation, solidifying the NSE's dominance. This storyline shows that the NSE's primacy appears to be derived more from tradition than from any demonstration of technical superiority. To date, path dependence is visible in on-going research efforts resulting from the popularity of the NSE. This historical perspective highlights how social processes have shaped the way hydrological models are evaluated.},
+  file = {C:\Users\TsyplenkovA\OneDrive - MWLR\ATS\Personal\zotero-library\Melsen et al.\Melsen et al._2025_The rise of the Nash-Sutcliffe efficiency in hydro.pdf}
+}
+
 @article{mizukamiChoiceCalibrationMetrics2019,
   title = {On the Choice of Calibration Metrics for ``High-Flow'' Estimation Using Hydrologic Models},
   author = {Mizukami, Naoki and Rakovec, Oldrich and Newman, Andrew J. and Clark, Martyn P. and Wood, Andrew W. and Gupta, Hoshin V. and Kumar, Rohini},
