+++ man/kge.Rd
@@ -104,7 +104,8 @@ Kling–Gupta efficiency scores. Hydrology and Earth System Sciences, 23,
 }
 \seealso{
 Other KGE variants: 
-\code{\link{kge2012}()}
+\code{\link{kge2012}()},
+\code{\link{kgelog_low}()}
 }
 \concept{KGE variants}
 \keyword{gof}
