+++ NAMESPACE
@@ -9,6 +9,7 @@ S3method(format, measure)
 export(nse)
 export(kge)
 export(kge2012)
+export(kgelog_low)
 export(mse)
 export(rmse)
 export(pbias)
@@ -21,6 +22,7 @@ export(gm)
 S3method(nse, data.frame)
 S3method(kge, data.frame)
 S3method(kge2012, data.frame)
+S3method(kgelog_low, data.frame)
 S3method(mse, data.frame)
 S3method(rmse, data.frame)
 S3method(pbias, data.frame)
@@ -33,6 +35,7 @@ S3method(gm, data.frame)
 export(nse_vec)
 export(kge_vec)
 export(kge2012_vec)
+export(kgelog_low_vec)
 export(mse_vec)
 export(rmse_vec)
 export(pbias_vec)
