### 2025-08-25 470dfa5
feat: kgelog_hi
