+++ NAMESPACE
@@ -10,6 +10,7 @@ export(nse)
 export(kge)
 export(kge2012)
 export(kgelog_low)
+export(kgelog_hi)
 export(mse)
 export(rmse)
 export(pbias)
@@ -23,6 +24,7 @@ S3method(nse, data.frame)
 S3method(kge, data.frame)
 S3method(kge2012, data.frame)
 S3method(kgelog_low, data.frame)
+S3method(kgelog_hi, data.frame)
 S3method(mse, data.frame)
 S3method(rmse, data.frame)
 S3method(pbias, data.frame)
@@ -36,6 +38,7 @@ export(nse_vec)
 export(kge_vec)
 export(kge2012_vec)
 export(kgelog_low_vec)
+export(kgelog_hi_vec)
 export(mse_vec)
 export(rmse_vec)
 export(pbias_vec)
