### 2025-08-25 eb24387
feat: kgelog, kgelog_low and kgelog_hi
