+++ tests/testthat/test-system-crash.R
@@ -102,6 +102,63 @@ test_that("kge2012", {
   )
 })
 
+test_that("kgelog", {
+  skip_if_not_installed("quickcheck")
+
+  quickcheck::for_all(
+    obs = quickcheck::any_atomic(any_na = TRUE),
+    sim = quickcheck::any_atomic(any_na = TRUE),
+    na_flag = quickcheck::logical_(any_na = FALSE),
+    property = function(obs, sim, na_flag) {
+      suppressWarnings(
+        try(
+          kgelog_vec(truth = obs, estimate = sim, na_rm = na_flag),
+          silent = TRUE
+        )
+      )
+      expect_true(TRUE)
+    }
+  )
+})
+
+test_that("kgelog_low", {
+  skip_if_not_installed("quickcheck")
+
+  quickcheck::for_all(
+    obs = quickcheck::any_atomic(any_na = TRUE),
+    sim = quickcheck::any_atomic(any_na = TRUE),
+    na_flag = quickcheck::logical_(any_na = FALSE),
+    property = function(obs, sim, na_flag) {
+      suppressWarnings(
+        try(
+          kgelog_low_vec(truth = obs, estimate = sim, na_rm = na_flag),
+          silent = TRUE
+        )
+      )
+      expect_true(TRUE)
+    }
+  )
+})
+
+test_that("kgelog_hi", {
+  skip_if_not_installed("quickcheck")
+
+  quickcheck::for_all(
+    obs = quickcheck::any_atomic(any_na = TRUE),
+    sim = quickcheck::any_atomic(any_na = TRUE),
+    na_flag = quickcheck::logical_(any_na = FALSE),
+    property = function(obs, sim, na_flag) {
+      suppressWarnings(
+        try(
+          kgelog_hi_vec(truth = obs, estimate = sim, na_rm = na_flag),
+          silent = TRUE
+        )
+      )
+      expect_true(TRUE)
+    }
+  )
+})
+
 test_that("pbias", {
   skip_if_not_installed("quickcheck")
 
