+++ man/kge2012.Rd
@@ -109,7 +109,7 @@ Kling–Gupta efficiency scores. Hydrology and Earth System Sciences, 23,
 \seealso{
 Other KGE variants: 
 \code{\link{kge}()},
-\code{\link{kgelog_low}()}
+\code{\link{kgelog}()}
 }
 \concept{KGE variants}
 \keyword{gof}
