+++ NAMESPACE
@@ -9,6 +9,7 @@ S3method(format, measure)
 export(nse)
 export(kge)
 export(kge2012)
+export(kgelog)
 export(kgelog_low)
 export(kgelog_hi)
 export(mse)
@@ -23,6 +24,7 @@ export(gm)
 S3method(nse, data.frame)
 S3method(kge, data.frame)
 S3method(kge2012, data.frame)
+S3method(kgelog, data.frame)
 S3method(kgelog_low, data.frame)
 S3method(kgelog_hi, data.frame)
 S3method(mse, data.frame)
@@ -37,6 +39,7 @@ S3method(gm, data.frame)
 export(nse_vec)
 export(kge_vec)
 export(kge2012_vec)
+export(kgelog_vec)
 export(kgelog_low_vec)
 export(kgelog_hi_vec)
 export(mse_vec)

[…]
