### 2025-08-25 06425d8
docs: kgelog docs
