+++ R/kge.R
@@ -39,7 +39,7 @@
 #' to understand the nature of model performance and consider defining
 #' explicit benchmarks based on the study context.
 #'
-#' For further discussion, see Knoben et al. (2019), who caution against
+#' For further discussion, see *Knoben et al.* (2019), who caution against
 #' directly translating NSE-based interpretation thresholds to KGE.
 #'
 #' @family KGE variants
@@ -166,7 +166,7 @@ kge_vec <- function(
 #' to understand the nature of model performance and consider defining
 #' explicit benchmarks based on the study context.
 #'
-#' For further discussion, see Knoben et al. (2019), who caution against
+#' For further discussion, see *Knoben et al.* (2019), who caution against
 #' directly translating NSE-based interpretation thresholds to KGE.
 #'
 #' @family KGE variants
@@ -248,17 +248,67 @@ kge2012_vec <- function(
   kge_cpp(truth, estimate, na_rm = na_rm, version = "2012")
 }
 
-
 #' Log-transformed Modified Kling-Gupta Efficiency
 #' @rdname kgelog
 #' @keywords gof
 #'
 #' @description
-#' Calculate the modified Kling-Gupta Efficiency (*Kling et al., 2012*),
-#' aka \eqn{KGE'}. Dimensionless (from \eqn{-\infty} to 1).
-#' `kge2012()` assesses the accuracy of
-#' simulated data by considering correlation, bias, and variability relative
-#' to observed data.
+#' Calculate the modified Kling-Gupta Efficiency (*Kling et al., 2012*) on
+#' **log-transformed** data as proposed in *Mai* (2023),
+#' namely \eqn{KGE_{log}}, \eqn{KGE_{logQ_{low}}} and \eqn{KGE_{logQ_{hi}}}.
+#' All are dimensionless (from \eqn{-\infty} to 1).
+#'
+#' This metric is recommended for emphasising low flows. By transforming the
+#' discharge data logarithmically, it gives more weight to smaller flow
+#' values, which is important for understanding drought conditions or
+#' baseflow behaviour (see *Mai 2023*; *Mizukami et al., 2019*).
+#'
+#' @details
+#' While the `kgelog()` function proposes the log-transformed version of the
+#' [kge2012], functions such as `kgelog_low()` and `kgelog_hi()`
+#' also perform data subsetting according to conditions specified in
+#' *Mai* (2023).
+#'
+#' The metrics `kgelog_low()` and `kgelog_hi()` are then the \eqn{KGE'}
+#' of the log-transformed observed and simulated streamflow considering
+#' only low-flow and high-flow time steps, respectively.
+#'
+#' A data point is considered in the derivation of `kgelog_low()` if the
+#' observed streamflow (\eqn{\text{obs}}) for that time step satisfies
+#' the following conditions:
+#'
+#' \deqn{
+#' 0.0 < \text{obs} \le min(\text{obs}) + 0.05 \times
+#' (max(\text{obs}) - min(\text{obs}))
+#' }
+#'
+#' A data point is considered in the derivation of `kgelog_hi()` if the
+#' observed streamflow (\eqn{\text{obs}}) for that time step satisfies
+#' the following conditions:
+#'
+#' \deqn{
+#' \text{obs} > min(\text{obs}) + 0.05 \times
+#' (max(\text{obs}) - min(\text{obs}))
+#' }
+#'
+#' @note
+#' Please note that the decision if a time step is a low-flow or high-flow
+#' time step is solely based on the observations which means it is always
+#' the same time steps for a given basin and time period while being
[…]
