### 2025-08-26 a64de74
docs: readme and news update
