### 2025-08-27 6cabc93
tests: better parameter-based testing
