+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/available_metrics: available_metrics.html
   articles/benchmarks: benchmarks.html
   articles/tidyhydro: tidyhydro.html
-last_built: 2025-08-27T02:23Z
+last_built: 2025-08-27T03:58Z
 urls:
   reference: https://atsyplenkov.github.io/tidyhydro/reference
   article: https://atsyplenkov.github.io/tidyhydro/articles
