### 2026-05-03 #713 @maurolepore
For context, our hold policy:

> The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.
> — https://devguide.ropensci.org/policies#policiesreviewprocess

The hold here started in July 2025, so we're approaching the 1-year mark. Looking forward to resuming before then.

Thanks!
